implemented stock updation feature

+++ categories/views.py
@@ -23,19 +23,19 @@ def create(request):
 
 def list(request):
     context = {}
-    datas = Category.objects.all()
+    datas = Category.objects.filter(is_deleted=False)
 
     context['datas'] = datas
     return render(request, 'categories/list.html', context)
 
 
-def edit(request, pk):
+def edit(request, slug):
     context = {}
     try:
-        data = Category.objects.get(pk=pk)
+        data = Category.objects.get(slug=slug)
     except Category.DoesNotExist as e:
-        messages.error(request, e)
-        return redirect('category:list')
+        messages.error(request, str(e))
+        return redirect('categories:list')
 
     form = CategoryForm(request.POST or None, instance=data)
 
@@ -49,43 +49,58 @@ def edit(request, pk):
     return render(request, 'categories/create.html', context)
 
 
-def delete(request, pk):
+def delete(request, slug):
     context = {}
     if request.method == 'POST':
         try:
-            data = Category.objects.get()
+            data = Category.objects.get(slug=slug)
             if data.is_deleted:
+                products_count = data.products.count()
+                if products_count > 0:
+                    messages.error(request,
+                                   "Unable to delete record. There are some product related with this category.")
+                    return redirect('archive:category_archive')
+
                 data.delete()
+
+                messages.success(request, 'Permanently deleted category data.')
+                return redirect('archive:category_list')
             else:
                 data.is_deleted = True
                 data.save()
-        except Category.DoesNotExist as e:
-            messages.error(request, e)
-            return redirect('categories:list')
 
-        messages.success(request, 'Successfully deleted the category record.')
+                messages.success(request, 'Successfully deleted category record.')
+                return redirect('categories:list')
+        except Category.DoesNotExist as e:
+            messages.error(request, str(e))
+            return redirect('categories:list')
 
     context['next'] = reverse('categories:list')
     context['page_name'] = "Category"
     context['status'] = 'Delete'
+    context['message'] = "Do you want to delete this data? It may be lost permanently."
+
     return render(request, 'snippets/delete.html', context)
 
 
-def undo(request, pk):
+def undo(request, slug):
     context = {}
     if request.method == 'POST':
         try:
-            data = Category.objects.get()
+            data = Category.objects.get(slug=slug)
 
             data.is_deleted = False
             data.save()
+
+            messages.success(request, 'Successfully restored category data.')
+            return redirect('archive:category_archive')
         except Category.DoesNotExist as e:
-            messages.error(request, e)
-            return redirect('categories:list')
+            messages.error(request, str(e))
+            return redirect('archive:category_archive')
 
-        messages.success(request, 'Successfully restored the category record.')
-
-    context['next'] = reverse('categories:list')
+    context['next'] = reverse('archive:category_archive')
     context['page_name'] = "Category"
     context['status'] = 'Undo'
+    context['message'] = "Do you want to undo this data?"
+
     return render(request, 'snippets/delete.html', context)
